backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
aiohttp==3.12.11
# File processing and OCR
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0
aiofiles==23.2.1

# Text processing
nltk==3.8.1

# HTTP client (for AI services)
httpx==0.25.2
requests==2.31.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.0.3

# Environment and configuration
python-dotenv==1.0.0

# Authentication (if needed)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background tasks
celery==5.3.4
redis==5.0.1

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production server
gunicorn==21.2.0